requirements.txt
certifi==2017.11.5
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.7
django-allauth==0.34.0
django-bootstrap==0.2.4
django-crispy-forms==1.7.0
djangorestframework==3.7.3
idna==2.6
oauthlib==2.0.6
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
urllib3==1.22
gunicorn==19.7.1
